refactor(types): generate public types from src instead of hand-maintaining them
Move the hand-authored types/index.esm.d.ts into src/types.ts as
real TS source that controller.ts/element.ts import from, and let
the build emit dist/index.esm.d.ts via
"tsc -p tsconfig.json --emitDeclarationOnly" (build =
"rollup -c && tsc ..."), matching sankey/treemap. package.json's
types/exports.types/files now point at dist instead of a
hand-maintained types/ directory.

MatrixController's public type comes from the real class now
(parameterized DatasetController<'matrix', ...>), which is strictly
more precise than the old bare `DatasetController` alias - dropped,
same as treemap does for its own controller alias.

MatrixElement is a concrete, non-generic class, which would have
narrowed the old generic `MatrixElement<T, O>` interface used for
typing custom element props/options. To avoid that regression, the
generic interface is kept in src/types.ts and re-exported as
MatrixElementType, mirroring treemap's TreemapElementType.

Move types/tests to test/types (checked via
"tsc --noEmit -p test/types/" in the typecheck script) and have it
import from the "chartjs-chart-matrix" package self-reference
instead of a relative path, so it type-checks against the real
generated declarations.
