+++ src/index.esm.ts
@@ -1,2 +1,13 @@
+export type {
+  AnchorX,
+  AnchorY,
+  MatrixControllerDatasetOptions,
+  MatrixDataPoint,
+  MatrixElement as MatrixElementType,
+  MatrixOptions,
+  MatrixParsedDataPoint,
+  MatrixProps,
+} from './types.js'
+
 export { default as MatrixController } from './controller.js'
 export { default as MatrixElement } from './element.js'
