+++ src/index.ts
@@ -5,4 +5,15 @@ import MatrixElement from './element.js'
 
 Chart.register(MatrixController, MatrixElement)
 
+export type {
+  AnchorX,
+  AnchorY,
+  MatrixControllerDatasetOptions,
+  MatrixDataPoint,
+  MatrixElement as MatrixElementType,
+  MatrixOptions,
+  MatrixParsedDataPoint,
+  MatrixProps,
+} from './types.js'
+
 export { MatrixController, MatrixElement }
