+++ src/types.ts
@@ -1,23 +1,21 @@
 import type {
   BorderRadius,
   CartesianScaleTypeRegistry,
-  Chart,
-  ChartComponent,
   CommonElementOptions,
   CommonHoverOptions,
   ControllerDatasetOptions,
   CoreChartOptions,
-  DatasetController,
   Element,
   ScriptableAndArrayOptions,
   ScriptableContext,
   VisualElement,
 } from 'chart.js'
 
-type AnyObject = Record<string, unknown>
+export type AnyObject = Record<string, unknown>
 
 export type AnchorX = 'left' | 'center' | 'right' | 'start' | 'end'
 export type AnchorY = 'top' | 'center' | 'bottom' | 'start' | 'end'
+
 export interface MatrixOptions extends Omit<CommonElementOptions, 'borderWidth'> {
   borderRadius: number | BorderRadius
   borderWidth: number | { top?: number; right?: number; bottom?: number; left?: number }
@@ -26,6 +24,7 @@ export interface MatrixOptions extends Omit<CommonElementOptions, 'borderWidth'>
   width: number
   height: number
 }
+
 export interface MatrixControllerDatasetOptions
   extends ControllerDatasetOptions,
     ScriptableAndArrayOptions<MatrixOptions, ScriptableContext<'matrix'>>,
@@ -43,7 +42,7 @@ export interface MatrixParsedDataPoint {
 }
 
 declare module 'chart.js' {
-  export interface ChartTypeRegistry {
+  interface ChartTypeRegistry {
     matrix: {
       chartOptions: CoreChartOptions<'matrix'>
       datasetOptions: MatrixControllerDatasetOptions
@@ -63,19 +62,8 @@ export interface MatrixProps {
   options?: Partial<MatrixOptions>
 }
 
-export type MatrixController = DatasetController
-export declare const MatrixController: ChartComponent & {
-  prototype: MatrixController
-  new (chart: Chart, datasetIndex: number): MatrixController
-}
-
 export interface MatrixElement<
   T extends MatrixProps = MatrixProps,
   O extends MatrixOptions = MatrixOptions,
 > extends Element<T, O>,
     VisualElement {}
-
-export declare const MatrixElement: ChartComponent & {
-  prototype: MatrixElement
-  new (cfg: AnyObject): MatrixElement
-}
