+++ test/types/options.ts
@@ -1,6 +1,5 @@
 import { Chart } from 'chart.js'
-
-import { MatrixController, MatrixElement } from '../index.esm'
+import { MatrixController, MatrixElement } from 'chartjs-chart-matrix'
 
 Chart.register(MatrixController, MatrixElement)
